wrote a wrapper class for the environment
